+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/state/RepairedAt.java
@@ -14,6 +14,8 @@
  */
 package com.ericsson.bss.cassandra.ecchronos.core.repair.state;
 
+import java.util.Collection;
+
 /**
  * Utility class to determine collective repaired at information for {@link VnodeRepairStates}.
  *
@@ -79,10 +81,15 @@ public final class RepairedAt
 
     public static RepairedAt generate(VnodeRepairStates vnodeRepairStates)
     {
+        return RepairedAt.generate(vnodeRepairStates.getVnodeRepairStates());
+    }
+
+    public static RepairedAt generate(final Collection<VnodeRepairState> vnodeRepairStates)
+    {
         long minRepairedAt = Long.MAX_VALUE;
         long maxRepairedAt = Long.MIN_VALUE;
 
-        for (VnodeRepairState vnodeRepairState : vnodeRepairStates.getVnodeRepairStates())
+        for (VnodeRepairState vnodeRepairState : vnodeRepairStates)
         {
             long repairedAt = vnodeRepairState.lastRepairedAt();