+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TestDataCenterAndVnodeRepairResourceFactory.java
@@ -66,7 +66,7 @@ public class TestDataCenterAndVnodeRepairResourceFactory
     private ReplicaRepairGroup generateReplicaRepairGroup(Host... hosts)
     {
         LongTokenRange range = new LongTokenRange(1, 2);
-        return new ReplicaRepairGroup(ImmutableSet.copyOf(hosts), ImmutableList.of(range));
+        return new ReplicaRepairGroup(ImmutableSet.copyOf(hosts), ImmutableList.of(range), System.currentTimeMillis());
     }
 
     private Host mockHost(String dataCenter, UUID hostId)