+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TestDataCenterRepairResourceFactory.java
@@ -56,7 +56,7 @@ public class TestDataCenterRepairResourceFactory
     private ReplicaRepairGroup generateReplicaRepairGroup(Host... hosts)
     {
         LongTokenRange range = new LongTokenRange(1, 2);
-        return new ReplicaRepairGroup(ImmutableSet.copyOf(hosts), ImmutableList.of(range));
+        return new ReplicaRepairGroup(ImmutableSet.copyOf(hosts), ImmutableList.of(range), System.currentTimeMillis());
     }
 
     private Host mockHost(String dataCenter)