+++ CHANGES.md
@@ -5,7 +5,7 @@
 * Repairs not scheduled when statistics disabled - Issue #175
 
 ### Merged from 1.0
-
+* Fix repair job priority - Issue #515
 * Fix malformed IPv6 for JMX - Issue #306
 * Step karaf to 4.2.8
 * Improve Alarm logging - Issue #191